Adjusting paths
Most paths will now be prefixed with /registration/ to clearly mark which
handler they belong to. Home page is the exception, because it's still
easier to just load the home page and display the registration form there.
Otherwise we'd need an extra redirect, which I'd rather avoid.

Renaming some of the handler methods as well, for better overall
consistency. index($f3) didn't say much. Now it's clear that  it's the
form, and right after that there's the form_process($f3) which is clear to
be the method to process the form submission.