Pending Review status for registrations
Changing the whole model, instead of a flag for waiting list, there's a generic status field, that could be extended in the future.

This status will allow putting people on hold, like in our case, if they're from outside of Poland and not members of Toastmasters, which makes it likely that they're just applying to get a visa.