Transaction isolation
We may have both the code in Payment::confirmation() and Payment::status_receive() running at the same time and we want to avoid problems there with status updates on the transaction and registration form.

These levels should ensure that when we're updating a transaction, the SELECTs on it will wait, and on the other hand, when we're SELECTing, the UPDATE will wait.