p24_method may be absent from POST
The guys from P24 confirmed that normally the field should always be
there, but we just had an instance where it wasn't and anyway it's not
required to validate a transaction back with P24. So, we'll just skip it
if it's not there and can always backfill it from the P24 transaction log.