Returning negative seats remaining
This will allow us to see if we're already over limit or not yet. Required
because once we pass the limit, we'll still be able to switch some
registrations manually to be able to pay, going beyond the initial limit
and if we're not checking for that, we'd always be updating these manually
enabled registrations back to waiting-list.